working beta
There may be a bug, or a problem with my subscription client,
but the last message reporting the last run durtation does not
get picked up by my subscriber.  It may be a problem with the
client or an interruption to the wireless when the transistor
cuts power to the sensors. More testing is needed.