Changed the way strings are created for mqtt messages. Removed OTA funtions.
The OTA funtions were removed beacuse that method of OTA will only work with
an always on device. Another mothod of updating will have to be developed.
The preparatioin of data strings to be tranmitted over MQTT are now much cleaner,
and more efficient - however all of the MQTT handling is likely to be revised in
the near future. For privacy most of the user defined variable have been moved to
conf.h, this includes SSID, PASSWORD, and MQTT_IP.